quiz_questions_flask/requirements.txt
blinker==1.6.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
flask-sqlalchemy==3.0.3 ; python_version >= "3.11" and python_version < "4.0"
flask==2.3.2 ; python_version >= "3.11" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.11" and platform_machine == "aarch64" and python_version < "4.0" or python_version >= "3.11" and platform_machine == "ppc64le" and python_version < "4.0" or python_version >= "3.11" and platform_machine == "x86_64" and python_version < "4.0" or python_version >= "3.11" and platform_machine == "amd64" and python_version < "4.0" or python_version >= "3.11" and platform_machine == "AMD64" and python_version < "4.0" or python_version >= "3.11" and platform_machine == "win32" and python_version < "4.0" or python_version >= "3.11" and platform_machine == "WIN32" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==1.10.7 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.30.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.13 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.0.2 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==2.3.4 ; python_version >= "3.11" and python_version < "4.0"
